fix(url): move urls into constants.ts
