+++ frontend/src/app/api/auth/signin/route.ts
@@ -1,5 +1,5 @@
 import {cookies} from "next/headers";
-import {DJANGO_BACKEND_URL} from "@/lib/constants";
+import {DJANGO_BACKEND_URL} from "@/app/api/constants";
 
 export async function POST(req: Request) {
     try {
