+++ frontend/src/app/api/auth/signout/route.ts
@@ -1,4 +1,5 @@
 import {cookies} from "next/headers";
+import {BACKEND_URL} from "@/lib/constants";
 
 
 export async function GET(req: Request) {
@@ -6,11 +7,11 @@ export async function GET(req: Request) {
     // delete access and refresh tokens
     // redirect to /
     if (!req.headers.get("x-user")) {
-        return Response.redirect("http://localhost:3000/", 302);
+        return Response.redirect(BACKEND_URL, 302);
     }
     const cookieStore = await cookies();
     cookieStore.delete("access");
     cookieStore.delete("refresh");
-    return Response.redirect("http://localhost:3000/", 302);
+    return Response.redirect(BACKEND_URL, 302);
 }
