+++ frontend/src/app/api/constants.ts
@@ -1,0 +1,1 @@
+export const DJANGO_BACKEND_URL = 'http://localhost:8000'
