+++ frontend/src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const BACKEND_URL = 'http://localhost:3000'
